requirements.txt
playwright>=1.40.0
python-telegram-bot>=20.7
aiohttp>=3.9.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
sqlalchemy>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Optional: Install lxml for faster HTML parsing (requires system libs)
# If you skip this, BeautifulSoup will use built-in html.parser
# lxml>=4.9.0

